requirements.txt
amqp==5.2.0
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
flower==2.0.1
humanize==4.8.0
idna==3.4
kombu==5.3.3
prometheus-client==0.18.0
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
tornado==6.3.3
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.9
